feat: update ai query
