+++ .env.example
@@ -9,3 +9,5 @@ STRAPI_API_URL="http://localhost:1337"
 STRAPI_API_TOKEN="your_strapi_api_token"
 DATA_API_URL=
 DATA_API_TOKEN=
+AI_API_URL=
+AI_API_TOKEN=
