+++ package.json
@@ -50,6 +50,7 @@
     "react-responsive": "^10.0.0",
     "remix-utils": "^7.6.0",
     "tailwind-merge": "^3.2.0",
+    "uuid": "^11.1.0",
     "viem": "^2.19.4",
     "zod": "3.23.8"
   },
