+++ src/domain/ai/repository/client.ts
@@ -1,3 +1,4 @@
+import HttpClient from "@/clients/http/HttpClient";
 import { createOpenAI } from "@ai-sdk/openai";
 
 const openai = createOpenAI({
@@ -5,4 +6,11 @@ const openai = createOpenAI({
   apiKey: process.env.OPENAI_API_KEY,
 });
 
-export default openai;
+const httpClient = new HttpClient({
+  baseUrl: process.env.AI_API_URL,
+  headers: {
+    Authorization: `Bearer ${process.env.AI_API_TOKEN}`,
+  },
+});
+
+export { openai, httpClient };
